backend/requirements.txt
Flask==2.3.2
tensorflow==2.9.1
Pillow==9.5.0
numpy==1.21.6
flask-cors==3.0.10